Prepare for release of GCS Client v0.2 to Maven Central.
When building the pom to upload to Maven Central, I'll locally set the revision to 0.2 (no SNAPSHOT). The version number is set to 0.3-SNAPSHOT in this CL so that all future CLs (or external SVN revisions) are known to be post-0.2.



Revision created by MOE tool push_codebase.
MOE_MIGRATION=6481


git-svn-id: http://appengine-gcs-client.googlecode.com/svn/trunk@62 d537cd30-a800-f30c-cc8c-eaa9e1c31ce9