Implementing a subset of seek no-op operations for StreamingBuffer
Summary:
This change lets GCS interface with the boto library. boto has a method
that downloads files from S3, which can resume from where it left off if
the download has an intermittent faillure. This implements the subset of
seek operations needed to make GCS files work within boto's expectations

Test Plan:
Can't test this code in webapp unless this diff is in master for this
submodule. This creates a new method and touches no other, so it
shouldn't break any existing code.

Reviewers: mattfaus

Differential Revision: https://phabricator.khanacademy.org/D18607