Make token-caching non-transactional.
Summary:
Even if this is called from inside a transaction, we don't need this
put() to be in the transaction: caching a token is safe and correct
even if the outer transaction ends up being rolled back.

Hopefully fixes https://khanacademy.atlassian.net/browse/INFRA-799

Test Plan:
Fingers crossed.  Will deploy and check for
   cross-group transaction need to be explicitly specified
in the logs after this is out.

Reviewers: colin, benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D42427