Cache GCS service discovery data in instance memory as well as memcache
Summary:
This commit fixes the same problem as D43303, but in a third-party
library.  In particular, the GCS library stores the discovery document
in an ndb model, using ndb auto-caching, which stores the data in
memcache as well as in an a per-thread cache.  (Actually, by default, it
doesn't write to datastore at all!  Although that's not at issue here.)
That's not good enough for us: hitting memcache once per request that
makes a call to GCS makes for a hot key.

So, now, we cache the discovery document in memory too.  It's a bit of a
mess to do so, the way the library is written, but so it goes.  I didn't
try to make it perfect, since we want to get this out before tomorrow's
peak.

Test Plan:
```
cd python/test
python test_runner.py ~/khan/webapp/third_party/frankenserver ../src .
```
gives only the same failures as before.
(Also, will run `make test` on jenkins.)

Reviewers: #sre, csilvers, boris

Reviewed By: #sre, csilvers, boris

Subscribers: boris

Differential Revision: https://phabricator.khanacademy.org/D61623